+++ packages/components/src/message-box/index.ts
@@ -12,7 +12,7 @@ const installExtra = (app: App) => {
 }
 
 YhMessageBox.install = (app: App) => {
-  app.config.globalProperties.$msgbox = MessageBox
+  app.config.globalProperties.$msgbox = MessageBox as any
   installExtra(app)
 }