+++ packages/hooks/src/use-click-outside/index.ts
@@ -8,7 +8,7 @@ export function useClickOutside(
 ) {
   if (typeof window === 'undefined') return
 
-  const listener = (event: MouseEvent | TouchEvent) => {
+  const listener = (event: Event) => {
     const el = unref(target)
     if (!el) return
 
@@ -17,7 +17,7 @@ export function useClickOutside(
     const path = event.composedPath()
     if (path.includes(el)) return
 
-    handler(event)
+    handler(event as MouseEvent | TouchEvent)
   }
 
   // 监听 mousedown 和 touchstart 以更早地处理，并使用 capture 确保能捕获到即使子元素停止了冒泡的事件